+++ CHANGELOG.md
@@ -6,6 +6,33 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 The following changes are pending, and will be applied on the next major release:
 
+## Unreleased changes
+
+### Patch changes
+- Fixed #2339: Unnamed policies are now returned by `getResourcePolicyAll` if an optional argument 
+  `{ acceptBlankNodes: true }` is specified. This additional argument makes this a non-breaking change,
+  as the current type signature isn't changed.
+- `getThing` now supports Blank Node identifiers in addition to IRIs and skolems to refer to a subject.
+- `getThingAll(dataset, { allowacceptBlankNodes: true })` now returns all Blank Nodes
+  subjects in the Dataset, in particular including those part of a single chain of
+  predicates. For instance, given the following dataset:
+
+  ```
+  @prefix ex: <https://example.org/> .
+  @prefix foaf: <http://xmlns.com/foaf/0.1/> .
+
+  ex:camille
+    foaf:knows [
+      foaf:name "Dominique"@en ;
+    ] .
+  ;
+  ```
+
+  `getThingAll(dataset, { allowacceptBlankNodes: true })` would have previously returned
+  a single element for the Named Node (`ex:camille`), it will now also include a second
+  element for the Blank Node. Blank Node identifiers are by definition unstable and shouldn't
+  be relied upon beyond local resolution.
+
 ## [2.0.1]
 
 The following changes have been implemented but not released yet: